+++ NEWS.md
@@ -1,3 +1,5 @@
+# waldo (development version)
+
 # waldo 0.2.2
 
 * Handle S4 objects that have attributes that are not slots.