+++ NEWS.md
@@ -1,11 +1,11 @@
 # waldo (development version)
 
-* Logical vectors fall back to element-by-element comparison in more cases 
-  (#51).
-
 * `compare()` gains a new `max_diffs` argument that allows you to control
   the maximum number of differences shown. Set `max_diffs = Inf` to 
   see all differences (#49)
+
+* Logical vectors fall back to element-by-element comparison in more cases 
+  (#51).
 
 * Long-form diff no longer confuses additions and deletions (#52, @krlmlr).