+++ NEWS.md
@@ -1,4 +1,4 @@
-# waldo (development version)
+# waldo 0.2.3
 
 * `compare()` gains a new `max_diffs` argument that allows you to control
   the maximum number of differences shown. Set `max_diffs = Inf` to